% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sondavi_snapshots}
\alias{sondavi_snapshots}
\title{The snapshots you have recorded}
\usage{
sondavi_snapshots(con)
}
\arguments{
\item{con}{A connection from [sondavi_connect()].}
}
\description{
The snapshots you have recorded
}
